Move GetObjectKeyFromSequenceNumber
